+++ srcs/check_twins.c
@@ -6,15 +6,15 @@
 /*   By: cghanime <cghanime@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/24 21:37:43 by cghanime          #+#    #+#             */
-/*   Updated: 2019/11/08 03:40:22 by cghanime         ###   ########.fr       */
+/*   Updated: 2019/11/25 16:07:59 by cghanime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/push_swap.h"
 
-int		check_twins(t_ps *ps, int value)
+uint8_t		check_twins(t_ps *ps, int64_t value)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (i < ps->len_a)
